requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
numpy==2.0.2
websockets==13.0.1
httpx==0.27.2
ccxt==4.4.37
scikit-learn==1.5.2
joblib==1.4.2
